requirements.txt
pyserial
fastapi
pydantic